# openapi.yaml
openapi: 3.0.1
info:
  title: Document Management System API
  description: |
    A comprehensive Document Management System REST API with features including:

    - **Document Management**: Create, read, update, and delete documents
    - **Content Management**: Upload and download content with flexible storage options
    - **Versioning**: Major and minor version control for documents
    - **Content Renditions**: Primary and secondary content renditions with transformations
    - **Full-Text Search**: Lucene-powered search with fielded queries
    - **File Store Management**: Manage multiple file storage locations
    - **Content Transformations**: PDF-to-text and other content transformations

    ## Key Features

    ### Document Types
    - ARTICLE, MANUAL, REPORT, SPREADSHEET, PRESENTATION, IMAGE, VIDEO, AUDIO, OTHER

    ### Storage Options
    - Database storage for small files
    - File system storage for large files
    - Automatic migration between storage types

    ### Search Capabilities
    - Simple text search across all fields
    - Fielded search (name, description, keywords, tags, content)
    - Boolean operators (AND, OR)
    - Phrase search, wildcard search
    - Indexes all document metadata and indexable content

    ### Content Renditions
    - Primary renditions (original content)
    - Secondary renditions (transformed content)
    - Automatic text extraction from PDFs
    - Extensible transformer plugin framework
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/documents/{id}/versions/minor:
    post:
      tags:
      - Documents
      summary: Create minor version
      description: "Create a new minor version (e.g., 1.0 → 1.1)"
      operationId: createMinorVersion
      parameters:
      - name: id
        in: path
        description: Item ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        "201":
          description: Minor version created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
  /api/documents/{id}/versions/major:
    post:
      tags:
      - Documents
      summary: Create major version
      description: "Create a new major version (e.g., 1.0 → 2.0)"
      operationId: createMajorVersion
      parameters:
      - name: id
        in: path
        description: Item ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        "201":
          description: Major version created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
...
  /api/documents/{id}/versions/history:
    get:
      tags:
      - Documents
      summary: Get version history
      description: Retrieve complete version history for an item
      operationId: getVersionHistory
      parameters:
      - name: id
        in: path
        description: Item ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Item not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SysObjectVersionDTO'
        "200":
          description: Version history retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SysObjectVersionDTO'
...
components:
  schemas:
...
    ContentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contentType:
          type: string
        sysObjectId:
          type: integer
          format: int64
        fileStoreId:
          type: integer
          format: int64
        fileStoreName:
          type: string
        storagePath:
          type: string
        size:
          type: integer
          format: int64
        formattedSize:
          type: string
        storageType:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    DocumentDTO:
      required:
      - documentType
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        majorVersion:
          type: integer
          format: int32
        minorVersion:
          type: integer
          format: int32
        parentVersionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        contents:
          type: array
          items:
            $ref: '#/components/schemas/ContentDTO'
        description:
          type: string
        documentType:
          type: string
          enum:
          - ARTICLE
          - REPORT
          - CONTRACT
          - MANUAL
          - PRESENTATION
          - TRIP_REPORT
          - OTHER
        tags:
          uniqueItems: true
          type: array
          items:
            type: string
        keywords:
          type: string
...
    SysObjectVersionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        version:
          type: string
        majorVersion:
          type: integer
          format: int32
        minorVersion:
          type: integer
          format: int32
        parentVersionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        isLatestVersion:
          type: boolean
        hasChildVersions:
          type: boolean